refactor: streamline settings display by removing unused theme options
- Removed theme-related strings and display widgets from localization files and settings page
- Introduced new LocaleText and ThemeText widgets for better management of locale and theme display
- Updated app_preferences to reflect changes in widget structure and removed obsolete utility functions

This update simplifies the settings management by focusing on essential display elements.
