+++ apps/assets/i18n/en.i18n.json
@@ -3,9 +3,6 @@
   "settings": {
     "title": "Settings",
     "language": "Language",
-    "theme": "Theme",
-    "theme_system": "System",
-    "theme_light": "Light",
-    "theme_dark": "Dark"
+    "theme": "Theme"
   }
 }
