+++ apps/assets/i18n/ja.i18n.json
@@ -3,9 +3,6 @@
   "settings": {
     "title": "設定",
     "language": "言語",
-    "theme": "テーマ",
-    "theme_system": "システム",
-    "theme_light": "ライト",
-    "theme_dark": "ダーク"
+    "theme": "テーマ"
   }
 }
